refactor(kubernetes): Move shared logic out of V1 provider
KubernetesAtomicOperationConverterHelper is used by both the V1 and V2
providers, so should not live in the V1 directory tree; move it up to
the base level.

KubernetesApiClientConfig is only used by the V1 provider, so it should
be moved into the v1 directory tree.
