+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v1/deploy/converters/securitygroup/DeleteKubernetesSecurityGroupAtomicOperationConverter.groovy
@@ -17,7 +17,7 @@
 package com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.converters.securitygroup
 
 import com.netflix.spinnaker.clouddriver.kubernetes.KubernetesOperation
-import com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.converters.KubernetesAtomicOperationConverterHelper
+import com.netflix.spinnaker.clouddriver.kubernetes.deploy.converters.KubernetesAtomicOperationConverterHelper
 import com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.description.securitygroup.DeleteKubernetesSecurityGroupDescription
 import com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.ops.securitygroup.DeleteKubernetesSecurityGroupAtomicOperation
 import com.netflix.spinnaker.clouddriver.orchestration.AtomicOperation
